Remove slideshow and embed view options in the Metadata configuration, which reduces the number of checkboxes.
We're experiencing issues with the number of form parameters submitted by the Metadata configuration page for an exhibit. Removing the slideshow and embed options, which are not used at UVic anyway, is a step toward fixing these issues.

Note that no changes are made to #available_view_fields or to the CatalogController's blacklight configuration because 1) available_view_fields is used elsewhere in the application and we don't want to mess with that behavior (at least for now), and 2) we still want to view results in gallery and slideshow views (without seeing the options in the exhibit form).